Restructure into module package
