--- classes/__init__.py
@@ -1,3 +1,0 @@
-from classes.display import Display
-from classes.rule import Rule
-from classes.token import Token
