+++ sner/classes/rule.py
