+++ sner/scripts/export_atf.py
@@ -1,10 +1,10 @@
-from scripts import professions, utilities
+from sner.scripts import professions, utilities
 import codecs
 import csv
 import re
 import argparse
 import random
-from classes import Display
+from sner.classes import Display
 
 
 left_features =  []
