+++ sner/scripts/ner/contextualfromnames.py
