+++ sner/scripts/ner/namesfromrule.py
