+++ sner/scripts/ner/spellingfromnames.py
